+++ tencentcloud/dcdb/v20180411/errors.go
@@ -74,6 +74,9 @@ const (
 	// 实例安全组信息更新失败。
 	FAILEDOPERATION_UPDATEINSTANCEINFOFAILED = "FailedOperation.UpdateInstanceInfoFailed"
 
+	// 用户未实名认证。
+	FAILEDOPERATION_USERNOTAUTHED = "FailedOperation.UserNotAuthed"
+
 	// VPC更新路由失败。
 	FAILEDOPERATION_VPCUPDATEROUTEFAILED = "FailedOperation.VpcUpdateRouteFailed"