+++ tencentcloud/sms/v20210111/models.go
@@ -1014,6 +1014,10 @@ type PullSmsSendStatus struct {
 
 	// 用户接收短信状态描述。
 	Description *string `json:"Description,omitempty" name:"Description"`
+
+	// 用户的 session 内容。与请求中的 SessionContext 一致，默认为空，如需开通请联系 [腾讯云短信小助手](https://cloud.tencent.com/document/product/382/3773#.E6.8A.80.E6.9C.AF.E4.BA.A4.E6.B5.81)。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	SessionContext *string `json:"SessionContext,omitempty" name:"SessionContext"`
 }
 
 type PullSmsSendStatusByPhoneNumberRequest struct {